+++ src/dawn/net/transport/Transport.h
@@ -4,7 +4,7 @@
  */
 #pragma once
 
-#include "core/Option.h"
+#include "core/Collections.h"
 
 namespace dw {
 enum class ServerConnectionState { NotListening, Listening };
@@ -31,7 +31,7 @@ public:
     virtual void update(float dt) = 0;
     virtual void send(ClientId client, const byte* data, u32 length) = 0;
     virtual Option<ServerPacket> receive(ClientId client) = 0;
-    virtual int numConnections() const = 0;
+    virtual usize numConnections() const = 0;
 
     virtual ServerConnectionState connectionState() const = 0;
 };
