+++ src/dawn/renderer/SceneGraph.h
@@ -36,8 +36,8 @@ public:
     // Frames.
     Frame* addFrame(SystemNode* frame_node);
     void removeFrame(Frame* frame);
-    Frame* frame(int i);
-    uint frameCount() const;
+    Frame* frame(usize i);
+    usize frameCount() const;
 
     // Callbacks. All take delta time (dt) as first parameter.
     Function<void(float)> preRenderSceneCallback;
