+++ src/dawn/renderer/VertexBuffer.cpp
@@ -7,7 +7,7 @@
 #include "renderer/Renderer.h"
 
 namespace dw {
-VertexBuffer::VertexBuffer(Context* context, Memory data, uint vertex_count,
+VertexBuffer::VertexBuffer(Context* context, Memory data, usize vertex_count,
                            const rhi::VertexDecl& decl, rhi::BufferUsage usage)
     : Object{context}, vertex_count_{vertex_count} {
     handle_ = context->module<Renderer>()->rhi()->createVertexBuffer(std::move(data), decl, usage);
@@ -17,7 +17,7 @@ VertexBuffer::~VertexBuffer() {
     context_->module<Renderer>()->rhi()->deleteVertexBuffer(handle_);
 }
 
-void VertexBuffer::update(Memory data, uint vertex_count, uint offset) {
+void VertexBuffer::update(Memory data, usize vertex_count, usize offset) {
     vertex_count_ = vertex_count;
     context_->module<Renderer>()->rhi()->updateVertexBuffer(handle_, std::move(data), offset);
 }
@@ -30,7 +30,7 @@ rhi::VertexBufferHandle VertexBuffer::internalHandle() const {
     return handle_;
 }
 
-u32 VertexBuffer::vertexCount() const {
+usize VertexBuffer::vertexCount() const {
     return vertex_count_;
 }
 }  // namespace dw
